Start working on dependency injection
- Old files moved in zzz folder while I work on updating them
- I will handle merge conflicts dont worry
