+++ Neuro/DependencyInjection/IContextAcceptor.cs
@@ -1,0 +1,6 @@
+﻿namespace Neuro.DependencyInjection;
+
+public interface IContextAcceptor
+{
+    public IContextProvider Context { set; }
+}
