+++ Neuro/DependencyInjection/IContextProvider.cs
@@ -1,0 +1,8 @@
+﻿using Neuro.Vision;
+
+namespace Neuro.DependencyInjection;
+
+public interface IContextProvider
+{
+    public IVisionHandler VisionHandler { get; }
+}
