+++ Neuro/Vision/IVisionHandler.cs
@@ -1,0 +1,10 @@
+﻿using Neuro.DependencyInjection;
+
+namespace Neuro.Vision;
+
+public interface IVisionHandler : IContextAcceptor
+{
+    public void Update();
+
+    public void ReportFindings();
+}
