+++ Neuro/zzz/Patches/Minigame_Patches.cs
